+++ apps/linows/src-tauri/src/platform/linux/mod.rs
@@ -2,6 +2,7 @@ pub mod autostart;
 pub mod clipboard;
 pub mod fonts;
 pub mod gnome_ext;
+pub mod gpu;
 pub mod icons;
 pub mod process;
 pub mod sysinfo;
